[ks44team02] 식단 등록 완성, 식단 수정 진행중 Author: 장민관

+++ ks44team02/src/main/java/ks44team02/dto/GoodsInfoImage.java
@@ -5,7 +5,7 @@ public class GoodsInfoImage {
 	private String infoImageOriginalName;
 	private String infoImageNewName;
 	private String infoImagePath;
-	private int infoImageSize;
+	private long infoImageSize;
 	private int infoImageIsLocal;
 	public String getInfoImageIdx() {
 		return infoImageIdx;
@@ -31,10 +31,10 @@ public class GoodsInfoImage {
 	public void setInfoImagePath(String infoImagePath) {
 		this.infoImagePath = infoImagePath;
 	}
-	public int getInfoImageSize() {
+	public long getInfoImageSize() {
 		return infoImageSize;
 	}
-	public void setInfoImageSize(int infoImageSize) {
+	public void setInfoImageSize(long infoImageSize) {
 		this.infoImageSize = infoImageSize;
 	}
 	public int getInfoImageIsLocal() {
